Move model picker open state to its own store
- Split model picker visibility out of shortcut modifier state
- Update sidebar and provider picker to use the new store
